+++ spec/models/work_package_spec.rb
@@ -239,7 +239,6 @@ describe WorkPackage do
 
         it "should only attach the attachment when saving was successful" do
           raw_attachments = [double('attachment')]
-          attachment = FactoryGirl.build(:attachment)
 
           Attachment.should_not_receive(:attach_files)
 
@@ -316,13 +315,72 @@ describe WorkPackage do
 
         role = FactoryGirl.create :role, permissions: [:move_work_packages]
 
-        member = FactoryGirl.create(:member, user: valid_user, project: project, roles: [role])
+        FactoryGirl.create(:member, user: valid_user, project: project, roles: [role])
       end
 
       subject { WorkPackage.allowed_target_projects_on_move.count }
 
       it "sees all active projects" do
         should eq Project.active.count
+      end
+    end
+  end
+
+  describe :duration do
+    #TODO remove once only WP exists
+    [:issue, :planning_element].each do |subclass|
+
+      describe "for #{subclass}" do
+        let(:instance) { send(subclass) }
+
+        describe "w/ today as start date
+                  w/ tomorrow as due date" do
+          before do
+            instance.start_date = Date.today
+            instance.due_date = Date.today + 1.day
+          end
+
+          it "should have a duration of two" do
+            instance.duration.should == 2
+          end
+        end
+
+        describe "w/ today as start date
+                  w/ today as due date" do
+          before do
+            instance.start_date = Date.today
+            instance.due_date = Date.today
+          end
+
+          it "should have a duration of one" do
+            instance.duration.should == 1
+          end
+        end
+
+        describe "w/ today as start date
+                  w/o a due date" do
+          before do
+            instance.start_date = Date.today
+            instance.due_date = nil
+          end
+
+          it "should have a duration of one" do
+            instance.duration.should == 1
+          end
+        end
+
+        describe "w/o a start date
+                  w today as due date" do
+          before do
+            instance.start_date = nil
+            instance.due_date = Date.today
+          end
+
+          it "should have a duration of one" do
+            instance.duration.should == 1
+          end
+        end
+
       end
     end
   end