+++ doc/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+* `#825` Migrate Duration
 * `#1421` Adapt issue created/updated wording to apply to work packages
 * `#1813` Run Data Generator on old AAJ schema